requirements.txt
bibtexparser
pylatexenc
PyYaml